Lint fixes and prettiness.

+++ xml.c
@@ -1,4 +1,4 @@
-/* $Id: xml.c,v 1.5 2008/12/01 16:01:28 kristaps Exp $ */
+/* $Id: xml.c,v 1.6 2008/12/02 00:10:37 kristaps Exp $ */
 /*
  * Copyright (c) 2008 Kristaps Dzonsons <kristaps@kth.se>
  *
@@ -167,9 +167,9 @@ mbuf_putstring(struct md_xml *p, const char *buf)
 static int
 mbuf_nputstring(struct md_xml *p, const char *buf, size_t sz)
 {
-	size_t		 i;
+	int		 i;
 
-	for (i = 0; i < sz; i++) {
+	for (i = 0; i < (int)sz; i++) {
 		switch (buf[i]) {
 		case ('&'):
 			if ( ! md_buf_puts(p->mbuf, "&amp;", 5))
